requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiomail==0.0.2
aiosignal==1.3.1
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
drymail==0.0.6
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mistune==3.0.2
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.10.3
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.9.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.50
ultralytics-thop==2.0.13
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==1.0.3
websockets==14.1
win32_setctime==1.2.0
yarl==1.18.3
